+++ src/query/storages/fuse/src/operations/read/parquet_rows_fetcher.rs
@@ -139,7 +139,7 @@ impl<const BLOCKING_IO: bool> ParquetRowsFetcher<BLOCKING_IO> {
 
             let blocks = compact_segment_info.block_metas()?;
             let block_idx = block_idx_in_segment(blocks.len(), block as usize);
-            let block_meta = &blocks[block_idx as usize];
+            let block_meta = &blocks[block_idx];
             let part_info = FuseTable::projection_part(
                 block_meta,
                 &None,
